Refactor cli module structure
